Studio metric selector split into components
