+++ src/ducks/Studio/Sidebar/MetricSelector/index.module.scss
